Consolidate mesh output tests into tests/mesh.rs with glTF index-width coverage
